+++ src/core/input-tree/core/probability/value-object/chance-value.ts
@@ -1,0 +1,11 @@
+export class ChanceValue {
+  private _value: unknown;
+
+  constructor(value: unknown) {
+    this._value = value;
+  }
+
+  value() {
+    return this._value;
+  }
+}