+++ src/errors/index.ts
@@ -22,10 +22,18 @@ export class WrongPossibleNullDefinitionError extends ChacaError {
 }
 
 export class EmptySequentialValuesError extends ChacaError {
-  constructor(public readonly fieldRoute: string) {
+  constructor(readonly fieldRoute: string) {
     super(`There are no more sequential values for the field '${fieldRoute}'`);
 
     this.name = "ChacaError.EmptySequentialValuesError";
+  }
+}
+
+export class WrongProbabilityFieldDefinitionError extends ChacaError {
+  constructor(readonly fieldRoute: string, message: string) {
+    super(`On '${fieldRoute}' ${message}`);
+
+    this.name = "ChacaError.WrongProbabilityFieldDefinitionError";
   }
 }
 
@@ -38,7 +46,7 @@ export class PickFieldDefinitionError extends ChacaError {
 }
 
 export class TryRefANoKeyFieldError extends ChacaError {
-  constructor(public readonly fieldRoute: string) {
+  constructor(readonly fieldRoute: string) {
     super(
       `The field '${fieldRoute}' is not a key field, so you can't reference this one`,
     );
@@ -48,10 +56,7 @@ export class TryRefANoKeyFieldError extends ChacaError {
 }
 
 export class NotEnoughValuesForRefError extends ChacaError {
-  constructor(
-    public readonly refFieldRoute: string,
-    public readonly keyFieldRoute: string,
-  ) {
+  constructor(readonly refFieldRoute: string, readonly keyFieldRoute: string) {
     super(
       `Not enough values of '${keyFieldRoute}' for the ref field '${refFieldRoute}'`,
     );
@@ -68,17 +73,14 @@ export class CyclicAccessDataError extends ChacaError {
 }
 
 export class NotExistFieldError extends ChacaError {
-  constructor(
-    public readonly fieldRoute: string,
-    public readonly refFieldRoute: string,
-  ) {
+  constructor(readonly fieldRoute: string, readonly refFieldRoute: string) {
     super(`From '${fieldRoute}', The field '${refFieldRoute}' does not exists`);
     this.name = "ChacaError.NotExistFieldError";
   }
 }
 
 export class EmptyEnumValuesError extends ChacaError {
-  constructor(public readonly fieldRoute: string) {
+  constructor(readonly fieldRoute: string) {
     super(`There are no values for the enum field '${fieldRoute}'`);
     this.name = "ChacaError.EmptyEnumValuesError";
   }