+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 
 - Added exception `WrongPossibleNullDefinitionError` for errors in the definition of the parameter `possibleNull`
 - Added exception `WrongArrayDefinitionError` for errors in the definition of the parameter `isArray`
+- Added exception `WrongProbabilityFieldDefinitionError` for errors in the definition of `probability` fields
 - 🗑️ `chaca.multiGenerate` was removed
 - 🎉 The number of items to select in the `pick` field can be defined as a range of values ​​or a function that returns the number of values ​​to choose from.